app/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
pydantic==1.10.7
sqlalchemy==1.4.41
jose==3.2.0
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.26.0
pytest==7.3.1

